Fixed Halo Mode(Thanks Scott McBride!)
Fixed redundancies in the code(Doing calculations with 0, having a copy of a variable). Data is written to the motor only once per loop.  Fixed program never switching into halo mode. Replaced wheelAdjust with wheelGain in Quickturn statement. Fixed comparison error as well in the quickturn statement(> to <). Added safety measure in case the controlller ever disconnected by placing everything in an if statement that checks the current status of the controller. Deadband values now use a total of 10% instead of 20% which conflicted with comments on the source code.